Safari-611.1.10/LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin-redirect-blob.html

<!DOCTYPE HTML>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Verifying navigator.sendBeacon(Blob) non-CORS cross-origin redirect handling.");

if (window.testRunner) {
  //testRunner.dumpPingLoaderCallbacks();
}

const blob = new Blob(["Cross", "Origin"], {type: "text/plain;from-beacon=true"});
// The "simple" parameter is just for differentiating the URLs.
shouldBeTrue('navigator.sendBeacon("http://127.0.0.1:8080/navigation/resources/redirection-response.php?status=302&simple=true&target=/non-existent.php", blob);');
</script>
