Safari-609.1.11.1/LayoutTests/http/tests/blink/sendbeacon/beacon-cross-origin-redirect-blob-expected.txt

Verifying navigator.sendBeacon(Blob) non-CORS cross-origin redirect handling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("http://127.0.0.1:8080/navigation/resources/redirection-response.php?status=302&simple=true&target=/non-existent.php", blob); is true
PASS successfullyParsed is true

TEST COMPLETE
